Move E2E tests into own file
