+++ test/spec/browser-index.js
@@ -38,3 +38,4 @@ require("./helpers/puppeteer/reporter.js");
 require("./test-common.js");
 require("./test-browser-specific.js");
 require("./test-parallel-uploads.js");
+require("./test-end-to-end.js");
