+++ test/spec/node-index.js
@@ -1,3 +1,4 @@
 require("./test-common.js");
 require("./test-node-specific.js");
 require("./test-parallel-uploads.js");
+require("./test-end-to-end.js");
